+++ src/player/musicPlayer/sourceMethods.js
@@ -1,6 +1,7 @@
 import { audiusMethods } from './audiusMethods.js';
 import { deezerMethods } from './deezerMethods.js';
 import { soundcloudMethods } from './soundcloudMethods.js';
+import { spotifyMethods } from './spotifyMethods.js';
 import { trackFactoryMethods } from './trackFactoryMethods.js';
 import { urlResolverMethods } from './urlResolverMethods.js';
 
@@ -10,4 +11,5 @@ export const sourceMethods = {
   ...soundcloudMethods,
   ...deezerMethods,
   ...urlResolverMethods,
+  ...spotifyMethods,
 };